Validação no Delete

+++ backend/src/controllers/TimesController.js
@@ -20,18 +20,30 @@ module.exports = {
     },
     async delete(request, response) {
         const { id } = request.params;
+        const password = request.headers.authorization;
+
+        const user = await connection('times')
+        .where('id', id)
+        .select('password')
+        .first();
+
+        if (user.password !== password) {
+            return response.status(401).json({ error: 'Operation not permitted.'});
+        }
 
         await connection('times').where('id', id).delete();
 
         return response.status(204).send();
-    }
-    // // async update(request, response) {
-    // //     const { name, description, password } = request.body;
+    
+    },
+      async update(request, response) {
+          const { id } = request.params;
+          const { name, description, password } = request.body;
 
-    // //     await connection('times')
-    // //     .where('name', name, 'description', description, 'password', password)
-    // //     .update('name', 'description', 'password');
+          await connection('times')
+          .where('id', id, 'name', name, 'description', description, 'password', password)
+          .update('name', 'description', 'password').returning('*');
 
-    // //     return response.json();
-    // }   
+          return response.status(200).send();;
+     }   
 };
